Add test for `htb`.
Separate tests into a separate file.
